delete pop up

+++ src/Pages/ManageAllOrders/ManageAllOrders.js
@@ -1,36 +1,45 @@
 import React, { useEffect, useState } from 'react';
+import useAuth from '../../hooks/useAuth';
 const ManageAllOrders = () => {
     const [users, setUsers] = useState([])
+    const {user} = useAuth();
     useEffect(() => {
         fetch(`https://ancient-island-16836.herokuapp.com/users`)
             .then(res => res.json())
             .then(data => {
-                const value= data.filter(db=>db.email==='web.faisal.bd@gmail.com')
+                const value= data.filter(db=>db.email===user.email)
                 setUsers(value);
             });       
     }, [])
 
+
+
+
+    
     const handleDelete= id => {
-        const url=`https://ancient-island-16836.herokuapp.com/users/${id}`;
-        fetch(url, {
-            method:'DELETE'
-        })
-        .then(res=>{
-            
-                    alert('Deleted successfully');
-                    
-                
-        })
+
+        const proceed = window.confirm('Are you sure,You want to delete?')
+        if (proceed) {
+            const url = `https://ancient-island-16836.herokuapp.com/users/${id}`;
+            fetch(url, {
+                method: 'DELETE'
+            })
+                .then(res => res.json())
+                .then(data => {
+                    if (data.deletedCount > 0) {
+                        alert('deleted successfully')
+                        const remainingUsers = users.filter(user => user._id !== id);
+                        setUsers(remainingUsers)
+
+                    }
+                })
+        }
+
+
         
     }
-    // -----------------
-    // -----------------
-    // -----------------
-    // -----------------
-    // -----------------
-    // -----------------
-    // -----------------
-    // -----------------
+    
+    
     return (
         <div id="users">
             <h2 className="mt-5 text-info">My orders</h2>
